feat: Radar steps completed and built
The radar prediction and update steps have been completed based on the code challenges in the lessons. The project builds but has not been tested, so there will most likely be changes. However, most of the work is complete and the laser prediction and updating needs to be completed before testing can happen.