feat: Complete Lidar update step and test on simulation
The Lidar update step, which is almost identical to the radar update step but in a different measurement space, has been completed. Small changes were also made to the prediction and radar update steps while testing. The project achieves RMSE values of {px: 0.06, py: 0.08, vx: 0.32, vy: 0.24} when run in the simulation.